backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
supabase==2.10.0
httpx==0.27.2
dateparser==1.1.8
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
python-multipart==0.0.6
cors==1.0.1
python-socketio==5.10.0
python-engineio==4.8.0
aiofiles==23.2.1
fastapi-socketio==0.0.10
